Implement Iterator.size_hint on Standard.
This is easy to do, improves allocation efficiency in various cases and is
otherwise occasionally useful. (It can be used to check “are there any
hyphenation opportunities?” without needing to query it twice or collect
into a vector or peek two elements along or any hack like that. This could
be done with a separate method also, size_hint being rather obscure.)