Serialize hyphenation patterns as tries during build
This commit includes two significant changes:
1. Patterns are bundled in their plain TeX form, and serialized at build
   time; this is preliminary to user-chosen normalization.
2. Patterns are generated by applying a binary serialization strategy
   directly to pattern and exception *tries*.

Change 1 entails a significant increase in the time of first-build,
partly owing to the limitation that `build.rs` cannot be compiled in
release mode for debug builds. (Refer to
[cargo #2424](https://github.com/rust-lang/cargo/issues/2424).)

The alternative to build-time serialization would be to package patterns
in every normalization form, and then either bundle them as a whole, or
isolate each as an optional data-only dependency. The former would
unacceptably inflate package size (to roughly 40MB, as of writing); the
latter I perceive as a misuse of Cargo crates.

Change 2 offers modest improvements in deserialization times, at the
expense of size:
- Deserializing the interned trie is about ~20–25% faster than
  deserializing sequential JSON data and building the trie at runtime.
- The interned data is approximately 25% larger. This may improve in the
  future, e.g. by compressing shared prefixes or suffixes.