Switch KLPTrie to byte keys
This change improves hyphenation performance by 10–20%, at the cost
of a marginal increase in the size of serialized tries. (Again, key
compression would be beneficial.)

A secondary adjustment is that `opportunities` now adopts a more
conservative strategy when segmenting uppercase words that change size
when lower-cased for scoring.

Previously, if `opportunities` generated a word-break opportunity at an
illegal index – which is to say, outside a character boundary – it
would shift the opportunity to the next available boundary.

Such behavior only affected words with characters extraneous to the
language patterns' alphabet, and produced "best-effort" results that
were, by construction, loosely orthographic.

Shifting the opportunity to a safe index would be more laborious with
byte-based patterns; instead, we consider misplaced opportunities to be
invalid. This decision may be revisited.